+++ src/zks_wallet/mod.rs
@@ -1,0 +1,5 @@
+mod errors;
+pub use errors::ZKSWalletError;
+
+mod wallet;
+pub use wallet::ZKSWallet;
